Game_Data_API/requirements.txt
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
